requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
pymupdf==1.24.9
chromadb==0.5.5
pandas==2.2.2
pyngrok==7.2.0